refactor states and schema to common module
inspired by clean architecture
